New mode to show threads
This adds a new mode to show only threads by default. The view which
shows all mails has been disabled. The user can now navigate to thread
mails the same way before to view individual mails.

This needs the latest HEAD of hs-notmuch which provides additional
functions to integration with notmuch. While this targets mainly
issue #97, this also closes issue #99 since with the integration the now
main index shows the total amount of messages in a thread.

Issue: #97